fix(config): update checkpointer imports after runtime move
